app-producer/requirements.txt
requests
kafka-python==2.0.2
yfinance
numpy
requests